+++ components/common/Header/mobile/MobileHeader.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import MobileHeader from '@/components/common/Header/MobileHeader';
+import MobileHeader from '@/components/common/Header/mobile/MobileHeader';
 
 export default {
   component: MobileHeader,
@@ -10,7 +10,9 @@ const Template: ComponentStory<typeof MobileHeader> = (args) => <MobileHeader {.
 
 export const Basic = Template.bind({});
 Basic.args = {
-  userName: '박건영',
-  userImage: '',
+  user: {
+    name: '박건영',
+    id: '1',
+  },
 };
 Basic.storyName = '기본';
